Factor out tasks and command implementations
